+++ src/main/kotlin/io/github/jhspetersson/turtlecommander/ui/FileManagerToolWindowFactory.kt
@@ -219,6 +219,7 @@ class FileManagerToolWindowFactory : ToolWindowFactory, DumbAware {
             "TurtleCommander.MoveFiles",
             "TurtleCommander.CreateDirectory",
             "TurtleCommander.CreateFile",
+            "TurtleCommander.CreateSymlink",
             "TurtleCommander.DeleteFiles",
             "TurtleCommander.SwitchPanel",
             "TurtleCommander.GoToFirst",